type cast

+++ source/s3_client.c
@@ -1069,7 +1069,7 @@ static struct aws_s3_meta_request *s_s3_client_meta_request_factory_default(
                 if (content_length < part_size) {
                     /* When the content length is smaller than part size and larger than the threshold, we set one part
                      * with the whole length */
-                    part_size = content_length;
+                    part_size = (size_t)content_length;
                 }
 
                 uint32_t num_parts = (uint32_t)(content_length / part_size);
